feat: moving to typescript
In this commit a lot of changes happen on the surface level but only a few happen under the hood. Generally the biggest difference is that everything is now written in typescript. To make sure that nothing breaks, it also contains minor fixes to examples and tests to make the tests pass.

Note: it currently contains a package dependency to the type-interface @tradle/constants branch.